Revert "Module cmd fix (#3250)"
This reverts commit a1131011263c086042ec24748e518d6e2b16b5f8.
