+++ lib/spack/spack/operating_systems/cnl.py
@@ -25,10 +25,10 @@
 import re
 
 from spack.architecture import OperatingSystem
+from spack.util.executable import *
 import spack.spec
 from spack.util.multiproc import parmap
 import spack.compilers
-from spack.util.module_cmd import get_module_cmd
 
 
 class Cnl(OperatingSystem):
@@ -63,7 +63,8 @@ class Cnl(OperatingSystem):
             if not cmp_cls.PrgEnv_compiler:
                 tty.die('Must supply PrgEnv_compiler with PrgEnv')
 
-            modulecmd = get_module_cmd()
+            modulecmd = which('modulecmd')
+            modulecmd.add_default_arg('python')
 
             output = modulecmd(
                 'avail', cmp_cls.PrgEnv_compiler, output=str, error=str)
