+++ lib/spack/spack/spec.py
@@ -120,7 +120,7 @@ import spack.util.spack_yaml as syaml
 from llnl.util.filesystem import find_headers, find_libraries, is_exe
 from llnl.util.lang import *
 from llnl.util.tty.color import *
-from spack.util.module_cmd import get_path_from_module, load_module
+from spack.build_environment import get_path_from_module, load_module
 from spack.error import SpecError, UnsatisfiableSpecError
 from spack.provider_index import ProviderIndex
 from spack.util.crypto import prefix_bits
